feat: #21 공통 버튼 추가
- GroupDetail에 공통버튼 추가
