+++ presenter/src/main/java/com/mashup/presenter/ui/common/ChinChinCommonComponent.kt
@@ -1,18 +1,24 @@
 package com.mashup.presenter.ui.common
 
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
+import androidx.compose.material.OutlinedButton
 import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.constraintlayout.compose.ConstraintLayout
 import com.mashup.presenter.ui.theme.Black
+import com.mashup.presenter.ui.theme.Grey_400
 import com.mashup.presenter.ui.theme.Primary_2
 
 @Composable
@@ -61,5 +67,34 @@ fun ChinChinCommonText(title: String, highlightText: String, modifier: Modifier 
             color = Primary_2,
             fontSize = 18.sp,
         )
+    }
+}
+
+@Composable
+fun ChinChinCommonButton(
+    icon: Int,
+    buttonText: String,
+    modifier: Modifier = Modifier,
+    onButtonClick: () -> Unit = {},
+) {
+    OutlinedButton(
+        onClick = { onButtonClick() },
+        modifier = modifier.defaultMinSize(1.dp, 1.dp),
+        contentPadding = PaddingValues(horizontal = 8.dp, vertical = 11.dp),
+    ) {
+        Row(
+            verticalAlignment = Alignment.CenterVertically,
+        ) {
+            Icon(
+                painter = painterResource(icon),
+                contentDescription = "",
+                tint = Grey_400
+            )
+            Text(
+                text = buttonText,
+                color = Grey_400,
+                modifier = Modifier.padding(start = 5.dp)
+            )
+        }
     }
 }
