+++ presenter/src/main/res/drawable/icon_user_more1.xml
@@ -1,0 +1,24 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M16,21V19C16,17.939 15.579,16.922 14.828,16.172C14.078,15.421 13.061,15 12,15H5C3.939,15 2.922,15.421 2.172,16.172C1.421,16.922 1,17.939 1,19V21"
+      android:strokeLineJoin="round"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M8.5,11C10.709,11 12.5,9.209 12.5,7C12.5,4.791 10.709,3 8.5,3C6.291,3 4.5,4.791 4.5,7C4.5,9.209 6.291,11 8.5,11Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="2"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M20,7C20.552,7 21,7.448 21,8V10H23C23.552,10 24,10.448 24,11C24,11.552 23.552,12 23,12H21V14C21,14.552 20.552,15 20,15C19.448,15 19,14.552 19,14V12H17C16.448,12 16,11.552 16,11C16,10.448 16.448,10 17,10H19V8C19,7.448 19.448,7 20,7Z"
+      android:fillColor="#000000"
+      android:fillType="evenOdd"/>
+</vector>
